+++ src/test/php/lang/ast/syntax/php/unittest/GenericsTest.class.php
@@ -37,6 +37,15 @@ class GenericsTest extends EmittingTest {
   }
 
   #[Test]
+  public function implements_generic_interface() {
+    $i= $this->type('interface <T><E> { }');
+    $t= $this->type('class <T><E> implements '.$i->getName().'<E> { }');
+
+    $c= Primitive::$STRING;
+    Assert::equals([$c], $t->newGenericType([$c])->getInterfaces()[0]->genericArguments());
+  }
+
+  #[Test]
   public function new_creates_generic_types() {
     $r= $this->run('class <T><E> {
       public function run() {